created grocery received screen
