+++ lib/constants/router_names.dart
@@ -9,4 +9,5 @@ const String homeScreenRoute = 'home';
 const String profileSettingsScreenRoute = 'settings';
 const String taskDetailsScreenRoute = 'taskDetails';
 const String taskAcceptScreenRoute = 'taskAccept';
-const String taskAcceptedScreenRoute = 'taskAccepted';
+const String groceryPendingScreenRoute = 'groceryPending';
+const String groceryReceivedScreenRoute = 'groceryReceived';
